Reference

MimiMooreEtAlAgricultureImpacts.get_model — Function

MimiMooreEtAlAgricultureImpacts.get_model(gtap::String; 
    pulse::Bool=false,
    floor_on_damages::Bool = true,
    ceiling_on_benefits::Bool = false)

Return a Mimi model with one component, the Moore Agriculture component. The user must specify the gtap input parameter as one of ["AgMIP_AllDF", "AgMIP_NoNDF", "highDF", "lowDF", "midDF"], indicating which gtap damage function the component should use.

The model has a time dimension of 2000:10:2300, and the fund_regions are the same as the FUND model.

Population and income levels are set to values from the USG2 MERGE Optimistic scenario. Temperature is set to output from the DICE model. If the user specifies pulse=true, then temperature is set to output from the DICE model with a 1 GtC pulse of CO2 emissions in 2020.

If floor_on_damages = true, then the agricultural damages in each timestep will not be allowed to exceed 100% of the size of the agricultural sector in each region. If ceiling_on_benefits = true, then the agricultural benefits in each timestep will not be allowed to exceed 100% of the size of the agricultural sector in each region.

MimiMooreEtAlAgricultureImpacts.get_ag_scc — Function

MimiMooreEtAlAgricultureImpacts.getagscc(gtap::String; prtp::Float64 = 0.03, horizon::Int = defaulthorizon, floorondamages::Bool = true, ceilingonbenefits::Bool = false)

Return the Agricultural SCC for a pulse in 2020 DICE temperature series and constant pure rate of time preference discounting with the specified keyword argument prtp. Optional keyword argument horizon can specify the final year of marginal damages to be included in the SCC calculation, with a default year of 2300.

If floor_on_damages = true, then the agricultural damages in each timestep will not be allowed to exceed 100% of the size of the agricultural sector in each region. If ceiling_on_benefits = true, then the agricultural benefits in each timestep will not be allowed to exceed 100% of the size of the agricultural sector in each region.